requirements.txt
firebase-admin==2.17.0
Pillow==6.1.0

